Partitioned array hash table and new benchmarks+tests

+++ src/hash_functions/murmur_hash_32.h
@@ -1,31 +1,3 @@
-/*
- * The MIT License (MIT)
- *
- * Copyright (c) 2016 Vivek Shah (University of Copenhagen)
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
-
- * The above copyright notice and this permission notice shall be included in
- * all copies or substantial portions of the Software.
-
- * THE SOFTWARE dbindex IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
- * THE SOFTWARE.
- *
- * murmur_hash_32.h
- *
- *  Created on: May 3, 2016
- *  Author: Vivek Shah <bonii @ di.ku.dk>
- */
 #ifndef SRC_HASH_FUNCTIONS_MURMUR_HASH_32_H_
 #define SRC_HASH_FUNCTIONS_MURMUR_HASH_32_H_
 
@@ -82,7 +54,7 @@ namespace dbindex {
 
             const std::uint32_t * blocks = (const std::uint32_t *) (ukey + nblocks * 4);
 
-            for (std::size_t i = -nblocks; i; i++) {
+            for (std::int32_t i = -nblocks; i; i++) {
                 value_t k1 = blocks[i];
 
                 k1 *= c1;
